+++ otx/algorithms/detection/adapters/mmdet/nncf/__init__.py
@@ -1,41 +1,14 @@
-# Copyright (C) 2021 Intel Corporation
+# Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
 from .compression import (
-    check_nncf_is_enabled,
     get_nncf_config_from_meta,
-    get_nncf_metadata,
-    get_uncompressed_model,
-    is_checkpoint_nncf,
     wrap_nncf_model,
-    is_state_nncf,
-    AccuracyAwareLrUpdater
-)
-from .compression_hooks import CompressionHook, CheckpointHookBeforeTraining
-from .runners import AccuracyAwareRunner
-from .utils import (
-    get_nncf_version,
-    is_accuracy_aware_training_set,
-    is_in_nncf_tracing,
-    no_nncf_trace,
 )
 
 from .monkey_patch import *
 
 __all__ = [
-    'AccuracyAwareRunner',
-    'AccuracyAwareLrUpdater',
-    'CheckpointHookBeforeTraining',
-    'CompressionHook',
-    'check_nncf_is_enabled',
-    'get_nncf_config_from_meta',
-    'get_nncf_metadata',
-    'get_nncf_version',
-    'get_uncompressed_model',
-    'is_accuracy_aware_training_set',
-    'is_checkpoint_nncf',
-    'is_in_nncf_tracing',
-    'is_state_nncf',
-    'no_nncf_trace',
-    'wrap_nncf_model',
+    "get_nncf_config_from_meta",
+    "wrap_nncf_model",
 ]
