--- otx/algorithms/detection/adapters/mmdet/utils/__init__.py
@@ -1,22 +1,0 @@
-"""OTX Adapters - openvino.utils."""
-
-# Copyright (C) 2022 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-from .config_utils import (
-    cluster_anchors,
-    config_from_string,
-    patch_config,
-    prepare_for_testing,
-    prepare_for_training,
-    set_hyperparams,
-)
-
-__all__ = [
-    "cluster_anchors",
-    "config_from_string",
-    "patch_config",
-    "prepare_for_testing",
-    "prepare_for_training",
-    "set_hyperparams",
-]
